+++ src/main/java/no/entur/uttu/graphql/mappers/StopPointInJourneyPatternMapper.java
@@ -28,7 +28,7 @@ import no.entur.uttu.model.StopPointInJourneyPattern;
 import no.entur.uttu.repository.FlexibleStopPlaceRepository;
 import no.entur.uttu.repository.ProviderRepository;
 import no.entur.uttu.repository.generic.ProviderEntityRepository;
-import no.entur.uttu.stopplace.StopPlaceRegistry;
+import no.entur.uttu.stopplace.spi.StopPlaceRegistry;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
