+++ src/test/java/no/entur/uttu/UttuTestApp.java
@@ -15,10 +15,9 @@
 
 package no.entur.uttu;
 
-import no.entur.uttu.ext.entur.organisation.EnturLegacyOrganisationRegistry;
+import no.entur.uttu.ext.entur.stopplace.EnturMummuStopPlaceRegistry;
 import no.entur.uttu.repository.generic.ProviderEntityRepositoryImpl;
 import no.entur.uttu.security.UttuSecurityConfiguration;
-import no.entur.uttu.stopplace.DefaultStopPlaceRegistry;
 import org.entur.pubsub.base.config.GooglePubSubConfig;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -42,7 +41,7 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
     @ComponentScan.Filter(type = FilterType.ASSIGNABLE_TYPE, value = App.class),
     @ComponentScan.Filter(
       type = FilterType.ASSIGNABLE_TYPE,
-      value = DefaultStopPlaceRegistry.class
+      value = EnturMummuStopPlaceRegistry.class
     ),
   }
 )
