+++ src/test/java/no/entur/uttu/stubs/StopPlaceRegistryStub.java
@@ -19,8 +19,8 @@ import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Optional;
-import no.entur.uttu.config.NetexHttpMessageConverter;
-import no.entur.uttu.stopplace.StopPlaceRegistry;
+import no.entur.uttu.ext.entur.stopplace.NetexHttpMessageConverter;
+import no.entur.uttu.stopplace.spi.StopPlaceRegistry;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpInputMessage;
 import org.springframework.stereotype.Component;
